back/requirements.txt
Django==5.1.5
psycopg2-binary==2.9.9
djangorestframework == 3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
ruff==0.6.7
django-filter==24.3
django-cors-headers==4.4.0
pillow==10.4.0
django-silk==5.2.0
django-allauth==65.0.2
djangorestframework-simplejwt==5.3.1
requests==2.32.3
cryptography==44.0.1
oauthlib==3.2.2
requests-oauthlib==2.0.0
mutagen==1.47.0
pydub==0.25.1
openai==1.66.3
drf-extra-fields==3.7.0
boto3==1.35.49
stripe==11.1.1
gunicorn==23.0.0
celery==5.4.0
redis==5.2.0
sib-api-v3-sdk==7.6.0
django-redis==5.4.0
